+++ spec/Listener/CodeCoverageListenerSpec.php
@@ -51,7 +51,7 @@ class CodeCoverageListenerSpec extends ObjectBehavior
         $this->afterSuite($event);
     }
 
-    function it_should_output_text_report(
+    function it_should_color_output_text_report_by_default(
         \PHP_CodeCoverage $coverage,
         \PHP_CodeCoverage_Report_Text $text,
         SuiteEvent $event,
@@ -67,9 +67,35 @@ class CodeCoverageListenerSpec extends ObjectBehavior
         ));
 
         $io->isVerbose()->willReturn(false);
+        $io->isDecorated()->willReturn(true);
         $this->setIO($io);
 
         $text->process($coverage, true)->willReturn('report');
+        $io->writeln('report')->shouldBeCalled();
+
+        $this->afterSuite($event);
+    }
+
+    function it_should_not_color_output_text_report_unless_specified(
+        \PHP_CodeCoverage $coverage,
+        \PHP_CodeCoverage_Report_Text $text,
+        SuiteEvent $event,
+        IO $io
+    ) {
+        $reports = array(
+            'text' => $text
+        );
+
+        $this->beConstructedWith($coverage, $reports);
+        $this->setOptions(array(
+            'format' => 'text'
+        ));
+
+        $io->isVerbose()->willReturn(false);
+        $io->isDecorated()->willReturn(false);
+        $this->setIO($io);
+
+        $text->process($coverage, false)->willReturn('report');
         $io->writeln('report')->shouldBeCalled();
 
         $this->afterSuite($event);
@@ -144,12 +170,10 @@ class CodeCoverageListenerSpec extends ObjectBehavior
             'blacklist_files' => array('src/filtered_file')
         ));
 
-        $filter->addDirectoryToWhitelist('src')->willReturn(null);
-        $filter->removeDirectoryFromWhitelist('src/filter')->willReturn(null);
-        $filter->addDirectoryToBlacklist('src/filter')->willReturn(null);
-        $filter->addFileToWhitelist('src/filter/whilelisted_file')->willReturn(null);
-        $filter->removeFileFromWhitelist('src/filtered_file')->willReturn(null);
-        $filter->addFileToBlacklist('src/filtered_file')->willReturn(null);
+        $filter->addDirectoryToWhitelist('src')->shouldBeCalled();
+        $filter->removeDirectoryFromWhitelist('src/filter')->shouldBeCalled();
+        $filter->addFileToWhitelist('src/filter/whilelisted_file')->shouldBeCalled();
+        $filter->removeFileFromWhitelist('src/filtered_file')->shouldBeCalled();
 
         $this->beforeSuite($event);
     }